+++ sabi-webclient/src/main/java/de/bluewhale/sabi/webclient/apigateway/PlagueServiceImpl.java
@@ -7,15 +7,19 @@ package de.bluewhale.sabi.webclient.apigateway;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import de.bluewhale.sabi.api.Endpoint;
+import de.bluewhale.sabi.exception.AuthMessageCodes;
 import de.bluewhale.sabi.exception.BusinessException;
 import de.bluewhale.sabi.exception.CommonExceptionCodes;
+import de.bluewhale.sabi.exception.Message;
 import de.bluewhale.sabi.model.PlagueRecordTo;
 import de.bluewhale.sabi.model.PlagueStatusTo;
 import de.bluewhale.sabi.model.PlagueTo;
+import de.bluewhale.sabi.webclient.utils.RestHelper;
 import jakarta.inject.Named;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.*;
+import org.springframework.web.client.RestClientException;
+import org.springframework.web.client.RestTemplate;
 import org.springframework.web.context.annotation.RequestScope;
 
 import java.util.Arrays;
@@ -100,7 +104,42 @@ public class PlagueServiceImpl extends APIServiceImpl implements PlagueService {
 
     @Override
     public void save(PlagueRecordTo plagueRecord, String pJWTBackendAuthtoken) throws BusinessException {
-        // TODO STS (02.10.22): ImplMe or drop it
+
+        String addPlageRecordURI = sabiBackendUrl + Endpoint.PLAGUE_CENTER_SERVICE+"/record"; // POST here
+        String requestJson;
+        try {
+            requestJson = objectMapper.writeValueAsString(plagueRecord);
+        } catch (JsonProcessingException e) {
+            log.error("Couldn't convert plage record object to json: {}",plagueRecord);
+            e.printStackTrace();
+            throw new BusinessException(CommonExceptionCodes.INTERNAL_ERROR);
+        }
+
+        RestTemplate restTemplate = new RestTemplate();
+        ResponseEntity<String> responseEntity;
+        HttpHeaders headers = RestHelper.prepareAuthedHttpHeader(pJWTBackendAuthtoken);
+        HttpEntity<String> requestEntity = new HttpEntity<>(requestJson,headers);
+
+        try {
+            responseEntity = restTemplate.postForEntity(addPlageRecordURI, requestEntity, String.class);
+            renewBackendToken(responseEntity);
+        } catch (RestClientException e) {
+            log.error(String.format("Couldn't reach %s",addPlageRecordURI),e.getLocalizedMessage());
+            throw new BusinessException(CommonExceptionCodes.NETWORK_ERROR);
+        }
+
+        if (!responseEntity.getStatusCode().is2xxSuccessful()){
+            /* FIXME STS (05.10.22): Fix status code mapping to real backend API  */
+            if (responseEntity.getStatusCode().equals(HttpStatusCode.valueOf(500))) {
+                log.error("Could not create plague record: {}",plagueRecord.toString());
+                throw new BusinessException(CommonExceptionCodes.INTERNAL_ERROR);
+            }
+            if (responseEntity.getStatusCode().equals(HttpStatusCode.valueOf(401))) {
+                log.warn("Invalid Token when trying to add a plague record: {}",plagueRecord.toString());
+                throw new BusinessException(Message.error(AuthMessageCodes.TOKEN_EXPIRED));
+            }
+        }
+
     }
 
     @Override